rebase mega-squash

+++ ArduPlane/servos.cpp
@@ -804,10 +804,10 @@ void Plane::set_servos(void)
         }
     }
 
-    uint8_t override_pct;
+    float override_pct;
     if (g2.ice_control.throttle_override(override_pct)) {
         // the ICE controller wants to override the throttle for starting
-        SRV_Channels::set_output_scaled(SRV_Channel::k_throttle, override_pct);
+        SRV_Channels::set_output_scaled(SRV_Channel::k_throttle, (int16_t)override_pct);
     }
 
     // run output mixer and send values to the hal for output
